requirements.txt
 click>=8.0.0
typer>=0.9.0
gitpython>=3.1.0
anthropic>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0
pyyaml>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0